% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_elements.R
\name{add_ribbon}
\alias{add_ribbon}
\title{Insert horizontal field to revealjs slide (e.g. footer, header)}
\usage{
add_ribbon(height = 100, offset = 50, pad = 10, em = 0.8,
  other_list = NULL, ...)
}
\value{
this function returns <div> tag for the footer.
}
\description{
.
This function is usually used as inline in revealjs_presentation.
For now, the ribbon position is controled manually by offset argument.
Elements other than style (e.g. class) will be controlled by passing a list.
}
