% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{draw_grid}
\alias{draw_grid}
\title{Draw grids in the slide}
\usage{
draw_grid(eval = TRUE, width = 2)
}
\arguments{
\item{eval}{draw grids or not, default TRUE}

\item{width}{linewidth of grids, default 2}
}
\description{
.
This function draws 100x100px grids in the slide to place objects
}
