% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_elements.R
\name{generate_tag}
\alias{generate_tag}
\title{Generate complete html tag by merging elements}
\usage{
generate_tag(styles, others, tag = c("<div ", "<span ")[1])
}
\value{
this function returns chr (class='foo: bar;')
}
\description{
.
merge css elements to complete html tag.
this function replaces underscore (_) to hyphen (-) to use handle input easily as inline.
}
