% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_fragment.R
\name{hook_fragment_index}
\alias{hook_fragment_index}
\title{Knitr hook for arranging order of the evaluation of fragments}
\usage{
hook_fragment_index(before, options, envir)
}
\description{
This hook is used for adding a 'data-fragment-index' tag to a figure.
This hook is usually used along with \code{hook_fragment_style}.
}
\seealso{
hook_fragment_style
}
