% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_fragment.R
\name{hook_fragment_style}
\alias{hook_fragment_style}
\title{Knitr hook for styling the fragment appearance}
\usage{
hook_fragment_style(before, options, envir)
}
\description{
This hook is used for adding a 'class' tag to a figure.
This hook is usually used along with \code{hook_fragment_index}.
In order to close <div> tag correctly, \code{hook_fragment_style} must be passed before the \code{hook_fragment_index}.
}
\seealso{
hook_fragment_index
}
