% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamishibai.R
\name{kamishibai}
\alias{kamishibai}
\title{Show layered-figures incrementally}
\usage{
kamishibai(figure_list,
  div = "top:0px; width:800px; height:600px; margin:0 auto")
}
\value{
this function returns chunks which shows figures.
}
\description{
.
This function is usually used as inline in revealjs_presentation.
}
\examples{
\dontrun{
<!--start rmarkdown-->
---
  output: revealjs::revealjs_presentation
---

# h1 level

```{r}
library(knitr)
library(tidyverse)
library(yose)
yose::set_reveal_hooks()

p <-
  lapply(2:5, function(x){
    tibble(x = 1:5, y = x * 2, index = 1:5)[1:x, ] \%>\%
      ggplot(aes(x, y)) +
      geom_point() +
      geom_line() +
      lims(x = c(0, 5), y = c(0, 10))
  })
```

## h2 level

`r yose::kamishibai(p)`
<!--end rmarkdown-->
}
}
