% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_elements.R
\name{parse_other}
\alias{parse_other}
\title{Convert R list object to css element}
\usage{
parse_other(other_list)
}
\value{
this function returns chr (class='foo: bar;')
}
\description{
.
parameters other than style (e.g. class) are determined with this internal function.
}
