% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_elements.R
\name{parse_style}
\alias{parse_style}
\title{Convert R list object for css styling to css element}
\usage{
parse_style(style_list)
}
\value{
this function returns chr (style='foo: bar;')
}
\description{
.
styles are determined with this internal function.
}
