% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_fragment.R
\name{set_reveal_hooks}
\alias{set_reveal_hooks}
\title{Set hooks for knitr hooks for revealjs presentation}
\usage{
set_reveal_hooks()
}
\value{
this function does not return any object.
}
\description{
Chunk options \code{fragment_style} and \code{fragment_index} are
passed to 'class' and 'fragment-data-index' in html, respectively.
This function may usually be evaluated in the setup chunk.
}
