% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_elements.R
\name{start_footer}
\alias{start_footer}
\title{Insert footer field to revealjs slide}
\usage{
start_footer(height = "100px", body_heigth = "600px", color = "#555555",
  backgroundcolor = "#aaaaaa")
}
\value{
this function returns <div> tag for the footer.
}
\description{
.
This function is usually used as inline in revealjs_presentation.
For now, the top end of the footer is controled manually by body_height argument.
}
