% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_elements.R
\name{start_header}
\alias{start_header}
\title{Insert header field to revealjs slide}
\usage{
start_header(height = "100px", body_offset = "-50px", color = "#555555",
  backgroundcolor = "#aaaaaa")
}
\value{
this function returns <div> tag for the header.
}
\description{
.
This function is usually used as inline in revealjs_presentation.
}
