% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinify.R
\name{tinify}
\alias{tinify}
\title{Compress image with tinypng API}
\usage{
tinify(path)
}
\arguments{
\item{path}{file path to the image file}
}
\value{
this function returns <div> tag for the footer.
}
\description{
.
Compress JPG/PNG image via tinypng API.
API key should be stored as global options for security.
}
